docs: add args table to all simple docs pages (#2144)
closes #2124
