+++ .storybook/preview.js
@@ -37,6 +37,9 @@ addParameters({
   previewTabs: {
     'storybook/docs/panel': {
       hidden: true
+    },
+    canvas: {
+      hidden: true
     }
   },
   docs: {
