+++ stories/components/picker/picker.js
@@ -7,11 +7,21 @@
 
 import { html } from 'lit';
 import { withCodeEditor } from '../../../.storybook/addons/codeEditorAddon/codeAddon';
+import { defaultDocsPage } from '../../../.storybook/story-elements/defaultDocsPage';
 
 export default {
   title: 'Components / mgt-picker',
-  component: 'mgt-picker',
-  decorators: [withCodeEditor]
+  component: 'picker',
+  decorators: [withCodeEditor],
+  parameters: {
+    docs: {
+      page: defaultDocsPage,
+      source: {
+        code:
+          '<mgt-picker resource="me/todo/lists" scopes="tasks.read, tasks.readwrite" placeholder="Select a task list" key-name="displayName"></mgt-picker>'
+      }
+    }
+  }
 };
 
 export const picker = () => html`
