+++ stories/components/tasks/tasks.stories.js
@@ -6,24 +6,26 @@
  */
 
 import { html } from 'lit';
-import { withCodeEditor } from '../../.storybook/addons/codeEditorAddon/codeAddon';
+import { withCodeEditor } from '../../../.storybook/addons/codeEditorAddon/codeAddon';
 
 export default {
-  title: 'Components / mgt-todo',
-  component: 'todo',
-  decorators: [withCodeEditor]
+  title: 'Components / mgt-tasks',
+  component: 'tasks',
+  decorators: [withCodeEditor],
+  parameters: {
+    docs: {
+      page: defaultDocsPage,
+      source: { code: '<mgt-tasks></mgt-tasks>' }
+    }
+  }
 };
 
 export const tasks = () => html`
-  <mgt-todo></mgt-todo>
-`;
-
-export const ReadOnly = () => html`
-  <mgt-todo read-only></mgt-todo>
+  <mgt-tasks></mgt-tasks>
 `;
 
 export const darkTheme = () => html`
-  <mgt-todo class="mgt-dark"></mgt-todo>
+  <mgt-tasks class="mgt-dark"></mgt-tasks>
   <style>
     body {
       background-color: black;
