+++ stories/components/teamsChannelPicker/teamsChannelPicker.a.js
@@ -7,11 +7,20 @@
 
 import { html } from 'lit';
 import { withCodeEditor } from '../../../.storybook/addons/codeEditorAddon/codeAddon';
+import { defaultDocsPage } from '../../../.storybook/story-elements/defaultDocsPage';
 
 export default {
   title: 'Components / mgt-teams-channel-picker',
   component: 'teams-channel-picker',
-  decorators: [withCodeEditor]
+  decorators: [withCodeEditor],
+  parameters: {
+    docs: {
+      page: defaultDocsPage,
+      source: {
+        code: '<mgt-teams-channel-picker></mgt-teams-channel-picker>'
+      }
+    }
+  }
 };
 
 export const teamsChannelPicker = () => html`
